refactor(directive): extract lifecycle-sync helpers (DIR-F7)
Move the three directive lifecycle-reconciliation helpers —
`syncPersistOptIn`, `syncMultiTabOptOut`, `syncElementRegistration`
— out of `directive.ts` (~140 LOC) into a focused
`directive-lifecycle.ts` module. The cluster reads `isRegisterValue`
from `directive.ts` (function declaration, ESM-hoist-safe in the
circular-import direction) and the persistence registry helpers
(`getOrAssignElementId`, `allowSensitivePersist`) from their existing
homes.

These helpers fire on every `created` / `beforeUpdate` hook across
every directive variant — text, checkbox, radio, select, the dynamic
gate. Co-locating them in their own file makes the bookkeeping
visible as a single cohesive unit (persist opt-in, multi-tab opt-out,
element-to-path registration) rather than three functions buried mid-
directive among the variant implementations.

`directive.ts` drops the now-unused `allowSensitivePersist` import.
Behavior-neutral: persistence, blank-persistence, multi-tab sync,
and v-register-component-runtime suites (83 tests) all stay green
unchanged.
